LayoutTests/fast/profiler/start-and-stop-profiling-in-the-same-function-expected.txt

CONSOLE MESSAGE: line 59: Profiling From A Nested Location
CONSOLE MESSAGE: line 75: Thread_1  0
CONSOLE MESSAGE: line 75:  functionWichStartsAndStopsTheProfiler start-and-stop-profiling-in-the-same-function.html 14
CONSOLE MESSAGE: line 75:  (idle)  0
This page's JavaScript starts profiling from within a nested location. 

To run this test manually, load it in the browser then load the WebInspector and look at the profile. In the profile there should be a call to functionWichStartsAndStopsTheProfiler().
